app/utils/io_control.py
import time

# 注：这是一个模拟实现，实际应用中需要根据硬件情况进行调整
# 在Windows系统上可以使用pyserial库控制串口
# 在Linux系统上可以使用GPIO库

try:
    # 尝试导入pyserial库
    import serial
    HAS_SERIAL = True
except ImportError:
    HAS_SERIAL = False

try:
    # 尝试导入RPi.GPIO库(树莓派)
    import RPi.GPIO as GPIO
    HAS_GPIO = True
    # 设置GPIO引脚
    GPIO_PIN = 17  # 使用GPIO17引脚输出信号
    GPIO.setmode(GPIO.BCM)
    GPIO.setup(GPIO_PIN, GPIO.OUT)
    GPIO.output(GPIO_PIN, GPIO.HIGH)  # 默认高电平
except ImportError:
    HAS_GPIO = False

# 串口配置
SERIAL_PORT = 'COM3'  # Windows系统通常为COMx，Linux系统通常为/dev/ttyUSBx
SERIAL_BAUDRATE = 9600

def send_low_signal():
    """
    发送低电平信号
    根据可用的硬件接口选择实现方式
    """
    if HAS_GPIO:
        # 使用GPIO发送低电平
        GPIO.output(GPIO_PIN, GPIO.LOW)
        print("发送低电平信号(GPIO)")
        time.sleep(0.5)  # 保持低电平0.5秒
        GPIO.output(GPIO_PIN, GPIO.HIGH)  # 恢复高电平
        return True
    
    elif HAS_SERIAL:
        # 使用串口发送信号
        try:
            ser = serial.Serial(SERIAL_PORT, SERIAL_BAUDRATE, timeout=1)
            ser.write(b'L')  # 发送'L'表示低电平
            print(f"通过串口{SERIAL_PORT}发送低电平信号")
            time.sleep(0.5)
            ser.close()
            return True
        except Exception as e:
            print(f"串口通信错误: {e}")
            return False
    
    else:
        # 如果没有可用硬件接口，则模拟发送
        print("模拟发送低电平信号(无硬件接口)")
        time.sleep(0.5)
        return True

def cleanup():
    """
    清理硬件资源
    """
    if HAS_GPIO:
        GPIO.cleanup()
        print("GPIO资源已清理")
    
    print("IO控制清理完成") 