requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
ultralytics>=8.0.0
torch>=1.7.0
numpy>=1.19.0
pyserial>=3.5 